requirements.txt
pyautogui
psutil
Pillow
screeninfo
setuptools